add command for throw the exception when data not found

+++ src/main/java/pws/a/learningRESTful/ProductServiceController.java
@@ -5,6 +5,7 @@
 package pws.a.learningRESTful;
 
 import exception.ProductAlreadyExist;
+import exception.ProductNotfound;
 import model.Product;
 import java.util.HashMap;
 import java.util.Map;
@@ -56,6 +57,7 @@ public class ProductServiceController
     @RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
     public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product)
     {
+        if(!productRepo.containsKey(id))throw new ProductNotfound();
         productRepo.remove(id);
         product.setId(id);
         productRepo.put(id, product);
